Move dynamic types to separate project #71
