+++ Source/QuantityTypes.sln
@@ -30,6 +30,10 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Csv", "Quanti
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Csv.Tests", "QuantityTypes.Csv.Tests\QuantityTypes.Csv.Tests.xproj", "{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Dynamic", "QuantityTypes.Dynamic\QuantityTypes.Dynamic.xproj", "{A3FEC30F-A820-4EA7-919C-7F9BE2B1686F}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes.Dynamic.Tests", "QuantityTypes.Dynamic.Tests\QuantityTypes.Dynamic.Tests.xproj", "{C2432EB1-94F6-4EDB-8197-68002B963C35}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -56,6 +60,14 @@ Global
 		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C4CD2F7B-9332-4BF8-9EED-0156BF95DF6F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A3FEC30F-A820-4EA7-919C-7F9BE2B1686F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A3FEC30F-A820-4EA7-919C-7F9BE2B1686F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A3FEC30F-A820-4EA7-919C-7F9BE2B1686F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A3FEC30F-A820-4EA7-919C-7F9BE2B1686F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2432EB1-94F6-4EDB-8197-68002B963C35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2432EB1-94F6-4EDB-8197-68002B963C35}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2432EB1-94F6-4EDB-8197-68002B963C35}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2432EB1-94F6-4EDB-8197-68002B963C35}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
